requirements.txt
bleach==2.1.3
certifi==2018.1.18
cffi==1.11.5
cycler==0.10.0
Cython==0.28
decorator==4.2.1
easydict==1.7
graphviz==0.8.2
h5py==2.7.1
html5lib==1.0.1
kiwisolver==1.0.1
matplotlib==2.2.0
munch==2.2.0
networkx==2.1
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.0.0
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.0.1
six==1.11.0
torch==0.3.1.post2
torchvision==0.2.0
tqdm==4.19.7
webencodings==0.5.1
wincertstore==0.2
